feat(auth): add password visibility toggle and prevent duplicate reset requests

+++ api/src/features/auth/auth.service.ts
@@ -470,6 +470,16 @@ export const resetPassword = async ({ email }: { email: string }) => {
 
   if (!user) throw new AppError("User not found", 404);
 
+  const notExpiredToken = await prisma.registerToken.findFirst({
+    where: { email, type: "RESET_PASSWORD", expiresAt: { gte: new Date() } },
+  });
+
+  if (notExpiredToken)
+    throw new AppError(
+      "You already have a reset password token, please check your email",
+      400,
+    );
+
   await prisma.registerToken.deleteMany({
     where: { email, type: "RESET_PASSWORD" },
   });
